fix code indentation, no functional changes
that GMC_cap_full indentation let me think it is a local variable
declaration (and thus, a bug) for a while, due to it being indented
with the variable declaration directly above it.

but it is not, it is correctly resetting the global flag.

also append the while() to the do-block, better understandable like
that.

also fix some minor de-alignments.