add TUBE_UNKNOWN 0 to have a specific value for experimenting
if one experiments with a not-yet-supported tube, one wants to
select the entry at index 0 in the tubes array (which is "Radiation
unknown" with a CPM to uSv/h conversion factor of 0.0).

Not sure whether sensor.info and other sites can deal yet with that
sensor type (if values are accidentally submitted), but I guess they
should ignore type 0 tube measurements.

Of course one should usually not submit measurements while
experimenting anyway.