add a basic, short README in English
this is about how a project README should look like, IMHO.

it should contain enough information for the reader to decide whether
this is the project they searched for.

but it should not be TL;DR (== too long; didn't read).

thus, it should NOT contain detailled installation and configuration
docs (that stuff belongs into the extensive docs the README links to).

this README has 2 uses / functions:
- being included as introduction into the long readthedocs.io based docs
- being shown on github (after review / community decision)

this avoids duplication and eases future changes/updates.