use bump2version tool for version bumps, fixes #169
derive everything from VERSION_STR, kill VERSION_DATE

bump2version can only deal with the string, not with dates.

the lora_software_version value is now computed from the string.

the date of a release can be looked up in the changelog (or in git).

for the version string, we display as much of it as possible
on the specific display (by just truncating it).
so if the display is wide enough, even the "release" part,
like "-dev" for development versions will be displayed or
partly displayed.

move version to version.h