rename lastMinuteLog*
usually last... is used in the code to remember a previous value,
e.g. of a count or a timestamp, to detect a difference or compute
and interval.

but here, "last" referred to the previous minute, which is also
implied by just minute_log* as it is after this change.