do not use test pin except when enabled, see #220
having a test pin that is toggled while the ISR for GM pulse counting
is active is nice for hardware debugging (e.g. with oscilloscope), but
not needed for production use.

by using a preprocessor #if block, the 2 digitalWrite() calls are not
done in production, which likely speeds up the ISR processing quite a
bit.