fix race condition, fixes #286
it could happen that display() was called before there was any GM event.
in that case, there were 2 timestamps initialized to millis() at different
times in program setup / subsystem setup time and dt could become
negative.

removed the millis() and just used 0 to initialize both.
check for the special condition and do not display/log/transmit anything
in that case.